Sort search tokens by cardinality to improve performance.
It seems that sqlite is unable to optimize queries by reordering WHERE
statement components for example. Therefore, we sort the search terms by
cardinality, so the (intermediate) result set is limited as much as
possible as early as possible.

If you search for "foo bar baz" for example, and "foo" yields the fewest
results, it will be sorted first in the query, etc.

In testing, this yielded huge performance improvements (few seconds if
sorted correctly vs. multiple minutes if sorted the wrong way round).

This also means it's no longer relevant in which order the user enters
the search terms on the command line.