Improved performance: token insert / index
Storing the max token id in the key_value table, we can remove the autoincrement
and therefore the index on token.id. Furthermore incrementing the token.id in
python allows us to insert the new tokens as bulk per file.

We can even display the new token count per file